Move simulation into the source
